requirements.txt
streamlit
groq
google-search-results
google-generativeai
requests
python-dotenv